move tests to the new file
